refactor: rename + move to right directories
